requirements.txt
imutils==0.5.2
language-selector==0.1
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
picamera==1.10
pickleshare==0.7.5
Pillow==3.1.2
python-dateutil==2.7.5
pytz==2018.7
requests==2.9.1
RPi.GPIO==0.6.5
scikit-learn==0.20.2
scipy==1.2.0
sklearn==0.0
sklearn-pandas==1.8.0
urllib3==1.13.1
wcwidth==0.1.7
xgboost==0.81

#sudo apt-get install python3-opencv

opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
python-debian==0.1.27
python-gnupg==0.3.8
python-systemd==231
usb-creator==0.3.0
xdiagnose==3.8.4.1

